+++ src/io/input.c
@@ -23,10 +23,10 @@ input_p read_input(FILE *f) {
 		exit(EXIT_FAILURE);
 	}
 	
-	for(int i = 0; i < in->rows; i++) {
+	for(size_t i = 0; i < in->rows; i++) {
 		in->data[i] = malloc(sizeof(double) * in->row_length);
 		if(in->data[i] == NULL) {
-                	for(int j = 0; j < i; j++) {
+                	for(size_t j = 0; j < i; j++) {
 				free(in->data[j]);
 			}
 			free(in->data);
@@ -36,11 +36,11 @@ input_p read_input(FILE *f) {
         	}
 	}
 
-	for(int i = 0; i < in->rows; i++) {
-		for(int j = 0; j < in->row_length; j++) {
+	for(size_t i = 0; i < in->rows; i++) {
+		for(size_t j = 0; j < in->row_length; j++) {
 			if( fscanf( f, "%lf", &(in->data[i][j]) ) != 1 ) {
 				fprintf(stderr, "Failed to get input data");
-				for(int k = 0; k < in->rows; k++) {
+				for(size_t k = 0; k < in->rows; k++) {
                 			free(in->data[k]);
 				}
 				free(in->data);
@@ -57,7 +57,7 @@ input_p read_input(FILE *f) {
 void free_input(input_p in) {
 	if(in != NULL) {
 		if(in->data != NULL) {
-			for(int i = 0; i < in->rows; i++) {
+			for(size_t i = 0; i < in->rows; i++) {
         			if(in->data[i] != NULL) {
 					free(in->data[i]);
 				}
