+++ src/loss.c
@@ -4,7 +4,7 @@
 // loss function: mean square error
 value_p mean_sqr_error(value_p* pred, value_p* actual, size_t num) {
 	value_p loss = make_value(0.0, 0.0, NULL, NULL);
-	for(int i = 0; i < num; i++) {
+	for(size_t i = 0; i < num; i++) {
 		loss = add_value(loss, pow_value(sub_value(pred[i], actual[i]), 2));
 	}
 	return loss;
